refactor(war-events): split WarEventLogService into core, points sync, and history modules
